move visualize_topology functions to util.py
